+++ R/text_spec.R
@@ -186,7 +186,7 @@ print.text_spec <- function(x, ..., verbose = FALSE){
 }
 
 
-# Write Functions -------------------------------------------------------
+# Write Text Functions -------------------------------------------------------
 
 #' @description A function to output strings for plain text content
 #' @details Basic logic is to wrap any text to the available line width, 
@@ -194,7 +194,7 @@ print.text_spec <- function(x, ..., verbose = FALSE){
 #' @param rs The Report Spec
 #' @param cntnt The text content to output
 #' @param lpg_rows Last page rows.
-#' @import stringi
+
 #' @noRd
 create_text_pages_text <- function(rs, cntnt, lpg_rows) {
   
@@ -218,7 +218,10 @@ create_text_pages_text <- function(rs, cntnt, lpg_rows) {
   return(rws)
 }
 
+
+
 #' Create list of vectors of strings for each page 
+#' @import stringi
 #' @noRd
 get_text_body <- function(rs, txt, line_width, line_count, lpg_rows, 
                           content_blank_row) {
@@ -301,3 +304,123 @@ get_text_body <- function(rs, txt, line_width, line_count, lpg_rows,
   return(ret)
   
 }
+
+
+# Write RTF Functions -------------------------------------------------------
+
+#' @noRd
+create_text_pages_rtf <- function(rs, cntnt, lpg_rows, conv) {
+  
+  if (!"report_spec" %in% class(rs))
+    stop("Report spec expected for parameter rs")
+  
+  if (!"report_content" %in% class(cntnt))
+    stop("Report Content expected for parameter cntnt")
+  
+  txt <- cntnt$object
+  
+  w <- rs$content_size[["width"]] 
+  
+  
+  # If user supplies a width, override default
+  if (!is.null(txt$width))
+    w <- txt$width
+  
+  rws <- get_text_body_rtf(rs, txt, w, rs$body_line_count, 
+                           lpg_rows, cntnt$blank_row, conv)
+  
+  return(rws)
+  
+  
+}
+
+#' Create list of vectors of strings for each page 
+#' @import stringi
+#' @noRd
+get_text_body_rtf <- function(rs, txt, line_width, line_count, lpg_rows, 
+                              content_blank_row, conv) {
+
+  
+  # Get titles and footnotes
+  ttls <- get_titles_rtf(txt$titles, line_width, conv) 
+  ftnts <- get_footnotes_rtf(txt$footnotes, line_width, conv) 
+  ttl_hdr <- get_title_header_rtf(txt$title_hdr, line_width, conv)
+  
+
+  if (txt$align == "right") 
+    algn <- "\\qr"
+  else if (txt$align %in% c("center", "centre"))
+    algn <- "\\qc"
+  else 
+    algn <- "\\ql"
+  
+  s <- paste0( algn, " ", txt$text, "\\line\\pard")
+  
+  print(get_lines_rtf(txt$text, line_width, rs$font, rs$font_size, rs$units))
+  
+
+  # Add blank above content if requested
+  a <- NULL
+  if (content_blank_row %in% c("both", "above"))
+    a <- "\\line"
+  
+  
+  # Add blank below content if requested
+  b <- NULL
+  if (content_blank_row %in% c("both", "below"))
+    b <- "\\line"
+  
+  # Combine titles, blanks, body, and footnotes
+  rws <- c(a, ttls, ttl_hdr, s, ftnts, b)
+  
+  ret <- rws
+  # Page list
+  #ret <- list()  
+  
+  # Create tmp variable for 1 page of content
+  tmp <- c()
+  
+  # # Offset the first page with remaining rows from the 
+  # # last page of the previous content
+  # offset <- lpg_rows 
+  # #print(paste("Offset:", offset))
+  # 
+  # # Assign content to pages
+  # for (i in seq_along(rws)) {
+  #   if (length(tmp) < (line_count - offset)) {
+  #     
+  #     # Append to existing page
+  #     tmp[length(tmp) + 1] <- rws[i]
+  #     
+  #   } else {
+  #     
+  #     # Start a new page
+  #     ret[[length(ret) + 1]] <- pad_any(tmp, line_width, 
+  #                                       get_justify(txt$align))
+  #     tmp <- rws[i]
+  #     
+  #     # Set to zero on second page and leave it that way
+  #     offset <- 0  
+  #   }
+  # }
+  # 
+  # # Deal with last page
+  # if (length(tmp) > 0 ) {
+  #   
+  #   
+  #   # If page is not empty
+  #   if (max(nchar(trimws(tmp))) > 0) {
+  #     
+  #     # Add last page
+  #     ret[[length(ret) + 1]] <- pad_any(tmp, line_width, 
+  #                                       get_justify(txt$align))
+  #   }
+  #   
+  # }
+  
+  
+  
+  return(ret)
+  
+}
+